Optimize reference resource handling
